feat: extract contents to /contents api and update docs
